Added LastUpdate field to Scenes and Sessions tables. This should  simplify garbage collection of scenes and sessions "post crash".
Added Name specification for all of the Scenes functions. Scene name
is a unique field so it should work as a key and it makes writing
administration tools a lot easier.

Split the enablement part of AddScene out into a separate interface.
This makes it easier to put disabled regions into the Scene db to,
for example, reserve regions.

Added a GetSessions interface. Just dumps the Sessions db.